requirements.txt
fastapi
uvicorn[standard]
jinja2
rich
pillow
httpx
